refactor: switch from local state of input search to url state
