+++ src/components/atoms/InputSearch/index.tsx
@@ -1,4 +1,4 @@
-import React, { useState } from 'react'
+import React, { ComponentProps } from 'react'
 import {
   ContainerInput,
   PlaceholderInput,
@@ -6,32 +6,12 @@ import {
   StyledSearchIcon,
 } from './styled'
 
-interface InputSearchProps {
-  onSearch: (query: string) => void
-}
+type InputSearchProps = ComponentProps<'input'>
 
-export default function InputSearch({ onSearch }: InputSearchProps) {
-  const [searchQuery, setSearchQuery] = useState('')
-
-  const handleInputChange = (e: React.ChangeEvent<HTMLInputElement>) => {
-    const inputValue = e.target.value
-    setSearchQuery(inputValue)
-
-    if (inputValue.length) {
-      onSearch(inputValue)
-    } else {
-      onSearch('')
-    }
-  }
-
+export default function InputSearch(props: InputSearchProps) {
   return (
     <ContainerInput>
-      <StyledInputSearch
-        type="text"
-        value={searchQuery}
-        onChange={handleInputChange}
-        id="input-search-mentors"
-      />
+      <StyledInputSearch type="text" {...props} id="input-search-mentors" />
       <PlaceholderInput htmlFor="input-search-mentors">
         Pesquise por nome
       </PlaceholderInput>
