--- src/components/molecules/MentorSubHeader/styled.ts
@@ -1,22 +1,0 @@
-import styled from 'styled-components'
-
-export const ContainerControls = styled.div`
-  width: 100%;
-  background-color: #fff;
-  padding: 2rem 0;
-  margin-top: 50px;
-`
-
-export const Content = styled.div`
-  display: flex;
-  justify-content: space-between;
-  padding: 0 2rem;
-  max-width: 1280px;
-  width: 100%;
-  margin: 0 auto;
-`
-
-export const ContainerSelects = styled.div`
-  display: flex;
-  gap: 10px;
-`
