+++ src/pages/mentores/index.tsx
@@ -1,52 +1,42 @@
 import { useState, useEffect } from 'react'
-import MentorSubHeader from '@/components/molecules/MentorSubHeader'
-import NoResult from '@/assets/noresult.svg'
-import Loading from '@/assets/loading.gif'
+
 import {
   Container,
-  MentorsContainer,
   SubHeaderContainer,
   TitleContainer,
   SubTitleContainer,
   CTAMain,
   CTASub,
-  NoResultContainer,
-  NoResultMain,
   StacksContainer,
   Stack,
   MainContent,
   Divider,
   RemoveFiltersBtn,
+  ContainerControls,
+  ContainerSelects,
+  Content,
 } from '@/styles/pages/mentors'
-import Image from 'next/image'
 import Link from 'next/link'
 import { Footer } from '@/components/molecules/Footer'
 import { useMentorsService } from '@/services/user/useMentorsService'
 import { IMentors } from '@/services/interfaces/IUseMentorsService'
-import dynamic from 'next/dynamic'
-const CardScheduling = dynamic(
-  () => import('@/components/atoms/CardSchedulingMentor'),
-  {
-    ssr: false,
-  },
-)
+import InputSearch from '@/components/atoms/InputSearch'
+import SelectFilter from '@/components/atoms/SelectFilter'
+import { genders, specialties } from '@/data/static-info'
+import { MentorsGrid } from '@/components/organisms/MentorsGrid'
+import { useRouter } from 'next/router'
 
 export default function MentorPage() {
+  const router = useRouter()
+
   const [genderFilter, setGenderFilter] = useState<string[]>([])
   const [specialtyFilter, setSpecialtyFilter] = useState<string[]>([])
-  const [mentorNameFilter, setMentorNameFilter] = useState('')
+  const [queryMentor, setQueryMentor] = useState('')
 
   const { fetchMentors, loading, mentors } = useMentorsService()
 
-  useEffect(() => {
-    const handleLoadFetchMentors = async () => {
-      await fetchMentors()
-    }
-    handleLoadFetchMentors()
-  }, [])
-
   const mentorsFiltered = mentors.filter((mentor: IMentors) => {
-    const nameFilter = mentorNameFilter.toLowerCase()
+    const nameFilter = queryMentor.toLowerCase()
 
     const hasSelectedSpecialty =
       specialtyFilter.length === 0 ||
@@ -63,8 +53,7 @@ export default function MentorPage() {
     return (
       hasSelectedGenders &&
       hasSelectedSpecialty &&
-      (!mentorNameFilter ||
-        mentor.fullName.toLowerCase().includes(nameFilter)) &&
+      (!queryMentor || mentor.fullName.toLowerCase().includes(nameFilter)) &&
       mentor.registerComplete === true
     )
   })
@@ -73,6 +62,29 @@ export default function MentorPage() {
     setSpecialtyFilter([])
     setGenderFilter([])
   }
+
+  useEffect(() => {
+    const handleLoadFetchMentors = async () => {
+      await fetchMentors()
+    }
+    handleLoadFetchMentors()
+  }, [])
+
+  useEffect(() => {
+    if (queryMentor) {
+      router.replace('/mentores', {
+        query: {
+          q: queryMentor,
+        },
+      })
+    }
+  }, [queryMentor])
+
+  useEffect(() => {
+    if (router.query.q) {
+      setQueryMentor(String(router.query.q))
+    }
+  }, [router.query.q])
 
   return (
     <>
@@ -91,15 +103,30 @@ export default function MentorPage() {
               </CTASub>
             </CTAMain>
           </SubHeaderContainer>
-          <MentorSubHeader
-            onGenderChange={(selectedOptions) =>
-              setGenderFilter(selectedOptions)
-            }
-            onSpecialtyChange={(selectedOptions) =>
-              setSpecialtyFilter(selectedOptions)
-            }
-            onMentorSearch={(query) => setMentorNameFilter(query)}
-          />
+
+          <ContainerControls>
+            <Content>
+              <InputSearch
+                value={queryMentor ?? ''}
+                onChange={(e) => setQueryMentor(e.target.value)}
+              />
+              <ContainerSelects>
+                <SelectFilter
+                  options={specialties}
+                  selectName="Especialidades"
+                  onChange={(option) => setSpecialtyFilter(option)}
+                  selectedOptions={specialtyFilter}
+                />
+                <SelectFilter
+                  options={genders}
+                  selectName="Gênero"
+                  onChange={(option) => setGenderFilter(option)}
+                  selectedOptions={genderFilter}
+                />
+              </ContainerSelects>
+            </Content>
+          </ContainerControls>
+
           {(specialtyFilter.length > 0 || genderFilter.length > 0) && (
             <StacksContainer>
               {specialtyFilter.map((selectedSpecialty) => (
@@ -114,30 +141,8 @@ export default function MentorPage() {
               </RemoveFiltersBtn>
             </StacksContainer>
           )}
-          <MentorsContainer>
-            {loading ? (
-              <>
-                <Image
-                  style={{ position: 'absolute', top: '10%', left: '45%' }}
-                  src={Loading}
-                  alt="Loading"
-                />
-              </>
-            ) : mentorsFiltered.length > 0 ? (
-              mentorsFiltered.map((mentor: IMentors) => (
-                <CardScheduling key={mentor.id} mentor={mentor} />
-              ))
-            ) : (
-              <NoResultContainer>
-                <Image src={NoResult} alt="Sem resultado" />
-                <NoResultMain>Nada por aqui!</NoResultMain>
-                <CTASub>
-                  Não conseguimos encontrar resultados pra sua busca.
-                </CTASub>
-                <CTASub>Tente alterar os filtros de pesquisa.</CTASub>
-              </NoResultContainer>
-            )}
-          </MentorsContainer>
+
+          <MentorsGrid loading={loading} mentors={mentorsFiltered} />
         </MainContent>
       </Container>
       <Footer />
